Updated build scripts and docs for Raspberry Pi

+++ base-vm/scripts/main-node/tools.sh
@@ -135,12 +135,15 @@ git clone --branch ${lensVersion} https://github.com/lensapp/lens.git
 cd ${INSTALLATION_WORKSPACE}/lens
 
 # Build OpenLens
-source /etc/profile.d/nvm.sh
-nvm use --delete-prefix lts/fermium
-make build || true # Do not fail KX.AS.CODE image build on error
-debOpenLensInstaller=$(find ${INSTALLATION_WORKSPACE}/lens/dist -name "OpenLens-*.deb")
-mv ${debOpenLensInstaller} ${INSTALLATION_WORKSPACE}
+if [[ -z $(which raspinfo) ]]; then
+  source /etc/profile.d/nvm.sh
+  nvm use --delete-prefix lts/fermium
+  make build || true # Do not fail KX.AS.CODE image build on error
+  debOpenLensInstaller=$(find ${INSTALLATION_WORKSPACE}/lens/dist -name "OpenLens-*.deb")
+  mv ${debOpenLensInstaller} ${INSTALLATION_WORKSPACE}
 
-# Tidy up
+  # Tidy up
+  sudo rm -rf ${INSTALLATION_WORKSPACE}/lens
+fi
+
 nvm use --delete-prefix lts/gallium
-sudo rm -rf ${INSTALLATION_WORKSPACE}/lens
